fix: iOS | Trigger "NotificationClickV2" (#71)
Update the iOS library to use the latest version that enables the correct triggering of the new NotificationClick event.
Update the implementation of the "FirebaseMessagingEventProtocol" to trigger multiple events simultaneously, including the new NotificationClick event.
Extend OSFCMClickableType and FirebaseNotificationType to implement the CustomStringConvertible protocol, applying the correct value to the description property.