fix(ios): remove returns by callback
Update the bridge methods so that it uses the new return mechanism done by the library. Doing this allows removing the 'OSCommonPluginLib' pod as it becomes useless.
The updates passes all the business logic, earlier done at library level, to the bridge.
Refactor the code a bit to make it simpler.

References: https://outsystemsrd.atlassian.net/browse/RMET-3554